Fix segmentation faults and major memory errors
First Error about stack memory:
Memory allocated on the stack cannot be addressed as heap memory once the function block has been deleted.

Second Error about memory leak:
As man strdup:
The strdup() function returns a pointer to a new string which is a duplicate of the string s. Memory for the new string is obtained with malloc(3), and can be freed with free(3).

This means we need to free up the memory.

Signed-off-by: Peppe289 <gsperanza204@protonmail.com>